+++ common/tests/wandb/test_wandb_context.py
@@ -6,10 +6,10 @@ import pytest
 import wandb
 from omegaconf import OmegaConf
 
-from metta.common.util.logging_helpers import setup_mettagrid_logger
 from metta.common.wandb.wandb_context import WandbConfigOff, WandbConfigOn, WandbContext
+from metta.util.init.logging import init_logging
 
-setup_mettagrid_logger()
+init_logging()
 logger = logging.getLogger("Test")
