+++ metta/util/init/logging.py
@@ -3,6 +3,7 @@ import os
 import sys
 from datetime import datetime
 
+import rich.traceback
 from rich.logging import RichHandler
 
 
@@ -67,7 +68,36 @@ def get_log_level(provided_level: str | None = None) -> str:
     return "INFO"
 
 
-def setup_mettagrid_logger(level: str | None = None) -> None:
+def init_file_logging(run_dir: str) -> None:
+    """Set up file logging in addition to stdout logging."""
+    # Create logs directory
+    logs_dir = os.path.join(run_dir, "logs")
+    os.makedirs(logs_dir, exist_ok=True)
+
+    node_index = os.environ.get("RANK", "0")
+    if node_index == "0":
+        log_file = "script.log"
+    else:
+        log_file = f"script_{node_index}.log"
+
+    # Set up file handler for the root logger
+    log_file = os.path.join(logs_dir, log_file)
+    file_handler = logging.FileHandler(log_file, mode="a")
+
+    # Use the same formatter as the existing console handler
+    formatter = logging.Formatter(fmt="%(asctime)s [%(levelname)s] %(name)s: %(message)s", datefmt="%H:%M:%S")
+    file_handler.setFormatter(formatter)
+    file_handler.setLevel(logging.INFO)  # Ensure file handler level is set
+
+    # Add to root logger so all log messages go to file
+    root_logger = logging.getLogger()
+    root_logger.addHandler(file_handler)
+
+    # Force a flush to make sure the file is created properly
+    file_handler.flush()
+
+
+def init_logging(level: str | None = None, run_dir: str | None = None) -> None:
     # Get the appropriate log level based on priority
     log_level = get_log_level(level)
 
@@ -88,3 +118,9 @@ def setup_mettagrid_logger(level: str | None = None) -> None:
     # set env COLUMNS if we are in a batch job
     if os.environ.get("AWS_BATCH_JOB_ID") or os.environ.get("SKYPILOT_TASK_ID"):
         os.environ["COLUMNS"] = "200"
+
+    rich.traceback.install(show_locals=False)
+
+    # Add file logging (after console handlers are set up)
+    if run_dir:
+        init_file_logging(run_dir)
