+++ tools/analyze.py
@@ -7,18 +7,18 @@ import hydra
 from omegaconf import DictConfig, OmegaConf
 
 from metta.agent.policy_store import PolicyStore
-from metta.common.util.logging_helpers import setup_mettagrid_logger
 from metta.eval.analysis import analyze
 from metta.eval.analysis_config import AnalysisConfig
-from metta.util.metta_script import setup_mettagrid_environment
+from metta.util.init.logging import init_logging
+from metta.util.init.mettagrid_environment import init_mettagrid_environment
 
 logger = logging.getLogger("analyze")
 
 
 @hydra.main(version_base=None, config_path="../configs", config_name="analyze_job")
 def main(cfg: DictConfig) -> None:
-    setup_mettagrid_environment(cfg)
-    setup_mettagrid_logger()
+    init_mettagrid_environment(cfg)
+    init_logging()
 
     logger.info(f"Analyze job config:\n{OmegaConf.to_yaml(cfg, resolve=True)}")
