+++ tools/map/gen.py
@@ -11,10 +11,10 @@ import hydra
 from omegaconf import DictConfig, OmegaConf
 
 from metta.common.util.config import config_from_path
-from metta.common.util.logging_helpers import setup_mettagrid_logger
 from metta.common.util.resolvers import register_resolvers
 from metta.map.utils.show import ShowMode, show_map
 from metta.map.utils.storable_map import StorableMap
+from metta.util.init.logging import init_logging
 
 # Aggressively exit on ctrl+c
 signal.signal(signal.SIGINT, lambda sig, frame: os._exit(0))
@@ -55,7 +55,7 @@ def uri_is_file(uri: str) -> bool:
 
 def main():
     register_resolvers()
-    setup_mettagrid_logger()
+    init_logging()
 
     parser = argparse.ArgumentParser()
     parser.add_argument("--output-uri", type=str, help="Output URI")
