+++ tools/map/view.py
@@ -3,11 +3,11 @@ import argparse
 import logging
 from typing import get_args
 
-from metta.common.util.logging_helpers import setup_mettagrid_logger
 from metta.common.util.resolvers import register_resolvers
 from metta.map.load_random import get_random_map_uri
 from metta.map.utils.show import ShowMode, show_map
 from metta.map.utils.storable_map import StorableMap
+from metta.util.init.logging import init_logging
 from tools.map.gen import uri_is_file
 
 logger = logging.getLogger(__name__)
@@ -15,7 +15,7 @@ logger = logging.getLogger(__name__)
 
 def main():
     register_resolvers()
-    setup_mettagrid_logger()
+    init_logging()
 
     parser = argparse.ArgumentParser()
     parser.add_argument(
